Move generic slice definition to its own file
Split SliceDefinition<TModel> out of SliceDefinition.cs into SliceDefinitionOfTModel.cs as requested in review.
