+++ src/RazorSlices/SliceDefinitionOfTModel.cs
@@ -1,0 +1,38 @@
+using System.Diagnostics.CodeAnalysis;
+
+namespace RazorSlices;
+
+/// <summary>
+/// Defines a strongly-typed Razor Slice that can be created via <see cref="RazorSliceFactory"/>.
+/// </summary>
+/// <typeparam name="TModel">The model type.</typeparam>
+/// <remarks>
+/// This type is used by the types generated by the Razor Slices source generator. You should not need to interact with this type directly in your project.
+/// </remarks>
+public sealed class SliceDefinition<TModel> : SliceDefinition
+{
+    /// <summary>
+    /// Creates a new instance of <see cref="SliceDefinition{TModel}"/>.
+    /// </summary>
+    /// <param name="sliceType">The type of the slice.</param>
+    public SliceDefinition(
+        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
+        Type sliceType)
+        : base(sliceType, RazorSliceFactory.GetSliceFactory<TModel>)
+    {
+        if (!HasModel || ModelType != typeof(TModel))
+        {
+            throw new InvalidOperationException($"""
+                Cannot create a strongly-typed slice definition for model type {typeof(TModel).Name} with slice {SliceType.Name}.
+                {(HasModel ? $"Ensure the model type is {ModelType!.Name}" : "It is not a strongly-typed slice.")}
+                """);
+        }
+    }
+
+    /// <summary>
+    /// Creates a new instance of the slice this definition represents with the specified model.
+    /// </summary>
+    /// <param name="model">The model for the slice.</param>
+    /// <returns>The slice instance.</returns>
+    public RazorSlice<TModel> CreateSlice(TModel model) => ((Func<TModel, RazorSlice<TModel>>)Factory)(model);
+}
